can save and load units that are spawned through the UI
Currently saving and loading is level agnostic.
The delete function does not reliable remove units from file due to normalization of 4-vectors in Stingray and rounding performed when saving to file.
Units are saved to file when spawned in through the UnitExplorer UI and removed from the save when deleted using the UnitExplorer  UI. Use "spawn_lvls" to spawn in saved units. To wipe the entire file use "clear_file". The save file is located in the executable directory as "level.txt"